refactor(color-select-panel): `color-new.ts` -> `color.ts`
